Fix compiler warning from implicit narrowing cast to uint8_t

+++ src/game.cpp
@@ -55,7 +55,10 @@ void SnakeGame::Reset() {
     // Reinitialise empty vector with snake part at the centre
     this->snake.clear();
     this->snake.resize(0);
-    this->snake.push_back({MapGridSizeHorizontal / 2, MapGridSizeVertical / 2});
+    this->snake.push_back({
+        static_cast<uint8_t>(MapGridSizeHorizontal / 2),
+        static_cast<uint8_t>(MapGridSizeVertical / 2)
+    });
 
     // Set starting snake length
     this->snakeLength = 4;
